Write class fields, including inherited fields
This is a draft for dumping the exact and complete layout of a C++ type
so that C can access the fields of said C++ type directly. This requires
working through vtables and virtual inheritance. This may not get the
alignment correct. Some static assertions are generated (for now) to try
and catch such issues in the future.