Force instantiation of templates directly
This implements what is hopefully a much better way to fill out the
missing class members and methods of specialized templates.

Before, it rams in some extra generated code at the end, which has the
issue that if yet more specializations were discovered while handling
the generated ones, they would not be handled and you'd have to use the
extra code hack again.

Now, the functions of the clang::Sema that are called when an explicit
instantiation is parsed, are called directly.

Additionally, the entire way that the DeclHandler handles forward
declarations, templates, and template specializations has been changed
to use queues. The writers in the queues are made to finish their work
until the queue is exhausted.