Allow anon enums to be given a name before forwarding
There are a few steps here. First, the enum writer is moved to a
separate file. Then the DeclWriter template has to allow skipping
identifier generation to avoid causing an exception. Then the enum
writer has to be modified to support a new case where the enum is
anonymous but was given a name elsewhere.

This allows a special case in identifier generation to be eliminated.