Eager template instantiation
This instantiates specialized templates as early as possible, so that
the generated struct can be used as a member of subsequent structs.

This requires the clang Sema object to be passed to the C++ class
writer. Also: the clang Sema object is used to generate the implicit
constructor, etc. So these don't have to be printed manually.

An issue where class members encountered by the AST consumer for some
unknown reason are printed before the class is typedef'd, has been
fixed, by changing the top level add method.

Some name printing function have been tweaked to give more useful output
in edge cases.