Fix non-deterministic Decl ordering, for real
Going from unordered_set to set was not correct because it sorts the
values, which are dynamically allocated pointers, which are not
deterministic. This goes for lists, and the implementation no longer
avoids adding duplicate pointers. However since all duplicates will be
removed when a dependency is satisfied, it shouldn't be an issue.