Clean up the CMake function for invoking unplusplus
Use more idiomatic CMake code: the argument parser, include guard.

Avoided the need for the unplusplus executable variable with a generator
expression that gets the target output.

Moved the list of standard C headers to be built in to the unplusplus
program. This list is (mostly) part of the C standard and should almost
always be used.